\name{lmeDGC-package}
\alias{lmeDGC-package}
\alias{lmeDGC}
\docType{package}
\title{lmeDGC}
\description{
}
\details{
\tabular{ll}{
Package: \tab gDGC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-03\cr
License: \tab GPL \cr
LazyLoad: \tab yes\cr
}


The package contains the following main functions: lmeDGC and MatrizDeDiferenciasDeMediasEstandarizadas.

}
\author{
	Julio A. Di Rienzo  <email: dirienzo@agro.unc.edu.ar>, Fernando Casanoves <email: casanoves@catie.ac.cr> and Maria del Carmen Romero <email: mariadelc.romero@gmail.com>
	
	Maintainer: 
	Julio A. Di Rienzo  <email: dirienzo@agro.unc.edu.ar> and Maria del Carmen Romero <email: mariadelc.romero@gmail.com>
 }
\references{

}
\keyword{ package }
\seealso{
	\code{\link{lmeDGC}},
	\code{\link{MatrizDeDiferenciasDeMediasEstandarizadas}}, 



}
\examples{

}
