\name{lmeDGC}
\alias{lmeDGC}
\title{lmeDGC}
\description{} 
\usage{
lmeDGC(MDE,Q,showplot,title)
}
\arguments{
  \item{MDE}{
}
  \item{Q}{
}
  \item{showplot}{

}
  \item{title}{
}

}
\details{

}
\value{


}
\references{
 
}
\author{
	Julio A. Di Rienzo  <email: dirienzo@agro.unc.edu.ar>
	Fernando Casanoves  <email: casanoves@catie.ac.cr>
       }
\note{

}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
	\code{\link{MatrizDeDiferenciasDeMediasEstandarizadas}}
}
\examples{
	lmeDGC(MDE,Q,showplot=TRUE,title="")

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
